requirements.txt
JPype1==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
Pillow==7.2.0
Werkzeug==	1.0.1
absl-py==	0.11.0
astor==	0.8.1
astunparse==	1.6.3
beautifultable==1.0.0
cachetools==	4.2.0
certifi==	2020.12.5
chardet==	4.0.0
cloudpickle==	1.6.0
flatbuffers==	1.12
future==	0.18.2
gast==0.3.3
google-auth==	1.24.0
google-auth-oauthlib==	0.4.2
google-pasta==	0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
numpy==	1.16.0
oauthlib==3.1.0
opt-einsum==3.3.0
pip	==20.3.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
requests==	2.25.1
requests-oauthlib==	1.3.0
rsa==4.6
scipy==	1.5.4
setuptools==50.3.2
six==1.15.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
wheel==	0.36.2
wrapt==	1.12.1
opencv-python==4.5.1.48